dicom/pyproject.toml
[tool.poetry]
name = "dicom"
version = "0.1.0"
description = ""
authors = ["Pablo Daniel Rey <pablo.rey@epicio.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
pydicom = "^2.4.4"
numpy = "^2.0.1"
pillow = "^10.4.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
